Extract comparison logic
